auth-service/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
pyjwt==2.3.0
passlib==1.7.4
argon2-cffi==23.1.0 